add ERNIE-DOC BPEtokenizer
