+++ paddlenlp/transformers/__init__.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 from .model_utils import PretrainedModel, register_base_model
-from .tokenizer_utils import PretrainedTokenizer
+from .tokenizer_utils import PretrainedTokenizer, BPETokenizer, tokenize_chinese_chars
 from .attention_utils import create_bigbird_rand_mask_idx_list
 
 from .bert.modeling import *
